Fix storing votes in database

+++ modules/database/models/vote.py
@@ -2,7 +2,7 @@ from modules.database.basemodel import *
 from .server import Server
 
 class Vote(BaseModel):
-    user_id = IntegerField(unique=True)
+    user_id = BigIntegerField(unique=True)
     server_id = ForeignKeyField(Server, 'id', backref='votes')
 
     class Meta:
